Fix ad hoc component generation
