+++ elm-tooling.json
@@ -1,0 +1,10 @@
+{
+    "entrypoints": [
+        "./src/Main.elm"
+    ],
+    "tools": {
+        "elm": "0.19.1",
+        "elm-format": "0.8.5",
+        "elm-json": "0.2.10"
+    }
+}
